add a button in the admin to no retry when truncate the errors table …

+++ Helper/Data.php
@@ -838,7 +838,7 @@ class Data extends \Magento\Framework\App\Helper\AbstractHelper
         $date = date('Y-m-d-H-i-s') . '-' . $msecArray[1];
         return $date;
     }
-    public function resetErrors($mailchimpStore)
+    public function resetErrors($mailchimpStore, $retry)
     {
         try {
             // clean the errors table
@@ -846,11 +846,11 @@ class Data extends \Magento\Framework\App\Helper\AbstractHelper
             $tableName = $this->_mailChimpErrors->getResource()->getMainTable();
             $connection->delete($tableName, "mailchimp_store_id = '".$mailchimpStore."'");
             // clean the syncecommerce table with errors
-            $connection = $this->_mailChimpSyncE->getResource()->getConnection();
-            $tableName = $this->_mailChimpSyncE->getResource()->getMainTable();
-            $connection->delete($tableName, "mailchimp_store_id = '".$mailchimpStore."' and mailchimp_sync_error is not null");
-//            $connection->commit();
-//            $connection->truncateTable($tableName);
+            if ($retry) {
+                $connection = $this->_mailChimpSyncE->getResource()->getConnection();
+                $tableName = $this->_mailChimpSyncE->getResource()->getMainTable();
+                $connection->delete($tableName, "mailchimp_store_id = '" . $mailchimpStore . "' and mailchimp_sync_error is not null");
+            }
         } catch (\Zend_Db_Exception $e) {
             throw new ValidatorException(__($e->getMessage()));
         }
